apps/videotransdub/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "videotransdub"
version = "0.2.0"
description = "Production-grade headless video translation/dubbing pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2,<3",
  "PyYAML>=6,<7",
]

[project.optional-dependencies]
# Real ASR engine (faster-whisper)
asr = [
  "faster-whisper>=1.0.0",
  "ctranslate2>=4.0.0",
]
# Free TTS engine (Edge-TTS)
tts = [
  "edge-tts>=6.1.0",
]
# Video inpainting (OpenCV)
inpaint = [
  "opencv-python-headless>=4.8.0",
]
# Streamlit UI
ui = [
  "streamlit>=1.30.0",
  "psutil>=5.9.0",
]
# Alibaba Qwen/DashScope translation
qwen = [
  "dashscope>=1.20.0",
]
# All production dependencies
full = [
  "videotransdub[asr,tts,inpaint,ui,qwen]",
]
# Colab-specific extras
colab = [
  "videotransdub[full]",
  "ipywidgets>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
